correction bits per long pour architectures 64 bits
(Logical change 1.1157)


git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@11261 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ pm2/pm2/marcel/include/asm-generic/linux_bitops.h
@@ -38,7 +38,7 @@
 
 #define ATOMIC_BITOPT_RETURN(op,retexpr) \
 { \
-	int	mask, old, new, ret; \
+	unsigned long	mask, old, new, ret; \
 	\
 	addr += nr / MA_BITS_PER_LONG; \
 	mask = 1 << (nr % MA_BITS_PER_LONG); \
@@ -53,7 +53,7 @@
 }
 #define BITOPT_RETURN(op,retexpr) \
 { \
-	int	mask, old; \
+	unsigned long	mask, old; \
 	\
 	addr += nr / MA_BITS_PER_LONG; \
 	mask = 1 << (nr % MA_BITS_PER_LONG); \
